+++ plugins/rime/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @livekit/agents-plugin-rime
 
+## 1.5.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @livekit/agents@1.5.2
+
+## 1.5.1
+
+### Patch Changes
+
+- Updated dependencies [[`5e8431a`](https://github.com/livekit/agents-js/commit/5e8431a7044d764bfcb575905068a7e13d39b5cf), [`8ded843`](https://github.com/livekit/agents-js/commit/8ded84301a2eb9de306b3a0aab2548726a488c17), [`213b284`](https://github.com/livekit/agents-js/commit/213b2840bb7b71146801258a8537f3ba28f633ff), [`c20907c`](https://github.com/livekit/agents-js/commit/c20907c048dfc07746b78a1d1d98fdc751c4a814), [`a0144bb`](https://github.com/livekit/agents-js/commit/a0144bb862f16f0c468ae529698bd8ac98adb00f), [`a0144bb`](https://github.com/livekit/agents-js/commit/a0144bb862f16f0c468ae529698bd8ac98adb00f), [`36e6460`](https://github.com/livekit/agents-js/commit/36e64603f62e41106b7712ce3c5a373bb516ffcd), [`92d7244`](https://github.com/livekit/agents-js/commit/92d7244b4fa1282b649f222c2459c6cfdf059067), [`9738429`](https://github.com/livekit/agents-js/commit/973842982c4377958dad3e462ae05683181dcdb4), [`1f881c2`](https://github.com/livekit/agents-js/commit/1f881c2b485b0dc0fcb20b835a470c7b243a4899), [`a0144bb`](https://github.com/livekit/agents-js/commit/a0144bb862f16f0c468ae529698bd8ac98adb00f), [`6dbf08f`](https://github.com/livekit/agents-js/commit/6dbf08fdb780ee2b841b117ba96103732fa4814f), [`7a5da25`](https://github.com/livekit/agents-js/commit/7a5da253691d00c1699cf39abe5a61aea22b8331), [`42a5355`](https://github.com/livekit/agents-js/commit/42a5355be4a53bdc4c9e251cef09fbb9002f1655)]:
+  - @livekit/agents@1.5.1
+
 ## 1.5.0
 
 ### Patch Changes